+++ PgJsonObjects/Synthetic/Ability/Ability Requirement/PgAbilityRequirementPetCount.cs
@@ -18,7 +18,7 @@
         }
 
         public double MaxCount { get { return RawMaxCount.HasValue ? RawMaxCount.Value : 0; } }
-        public double? RawMaxCount { get { return GetInt(4); } }
+        public double? RawMaxCount { get { return GetDouble(4); } }
         public RecipeKeyword PetTypeTag { get { return GetEnum<RecipeKeyword>(8); } }
     }
 }
