+++ PgJsonObjects/Synthetic/Recipe/PgRecipeCost.cs
@@ -18,7 +18,7 @@
         }
 
         public double Price { get { return RawPrice.HasValue ? RawPrice.Value : 0; } }
-        public double? RawPrice { get { return GetInt(0); } }
+        public double? RawPrice { get { return GetDouble(0); } }
         public RecipeCurrency Currency { get { return GetEnum<RecipeCurrency>(4); } }
     }
 }
